Adding 102-fibonacci.c file

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,15 +8,15 @@
 
 int main(void)
 {
-	int i, first = 0, second = 1, temp = 1;
+	long int i, first = 0, second = 1, temp = 1;
 
 	for (i = 1; i <= 50; i++)
 	{
 		temp = first + second;
 		if (i != 50)
-			printf("%d, ", temp);
+			printf("%ld, ", temp);
 		else
-			printf("%d", temp);
+			printf("%ld", temp);
 		first = second;
 		second = temp;
 	}
